scripts/firestore_backup_function/requirements.txt
google-cloud-firestore>=2.13.0
google-auth>=2.25.0
requests>=2.31.0
